v0.2.0 Movement & Lamp Rotation
The player can now move around, which was made possible by importing some of the moving and jumping functionality from my Unity 3D Test project with some modifications.

The lamp was added, which rotates around the player and faces the right direction. THIS WAS A PAIN TO GET PROPERLY IMPLEMENTED.