v0.2.1 Lamp Rotation Fix & Active Lamp
The Lamp now has a variable that activates and deactivates it for the purpose of player rotation. If the lamp is active, the player rotates to look at the horizontal direction to which it is pointing. Otherwise, it looks towards the direction of movement. This is the same as in the original game.

Some and foldersscripts were reorganized.